2008-09-16	Joel Sherrill <joel.sherrill@oarcorp.com>
	* shared/m68kbspgetworkarea.c: Add use of bsp_get_work_area() in its
	own file and rely on BSP Framework to perform more initialization.
	Remove unnecessary includes of rtems/libio.h and rtems/libcsupport.h.
	* shared/m68kpretaskinghook.c: Removed.

+++ c/src/lib/libbsp/m68k/shared/m68kbspgetworkarea.c
@@ -37,8 +37,8 @@ void bsp_get_work_area(
 )
 {
   *work_area_start       = _WorkspaceBase;
-  *work_area_size       = (unsigned long)_RamBase + (unsigned long) _RamSize -
-       (unsigned long)_WorkspaceBase;
+  *work_area_size       = (uintptr_t)_RamBase + (uintptr_t) _RamSize -
+       (uintptr_t)_WorkspaceBase;
   *heap_start = BSP_BOOTCARD_HEAP_USES_WORK_AREA;
   *heap_size = (size_t) _HeapSize;
 }
